pyproject.toml
[project]
name = "face-crop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dlib>=20.0.0",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
]
